third_party/blink/web_tests/virtual/at-property/fast/css/css-properties-values-api/at-property-cssom-expected.txt

This is a testharness.js-based test.
PASS Rule is applied
PASS Can read "syntax" from declaration
PASS Can read "inherits" from declaration
PASS Can read "initial-value" from declaration
Harness: the test ran to completion.
